norm: make_map.c, make_map_utiles.c
