+++ Makefile
@@ -3,6 +3,7 @@ SRCS = ./src/set_data.c \
 		./src/free_utils.c \
 		./src/main.c \
 		./src/make_map.c \
+		./src/make_map_utiles.c \
 		./src/draw.c \
 		./src/get_map_size.c \
 		./src/isometric.c \
